split off filetree from changes to reuse in stash
