+++ src/components/mod.rs
@@ -2,6 +2,7 @@ mod changes;
 mod command;
 mod commit;
 mod diff;
+mod filetree;
 mod help;
 mod msg;
 mod reset;
@@ -10,6 +11,7 @@ pub use changes::ChangesComponent;
 pub use command::{CommandInfo, CommandText};
 pub use commit::CommitComponent;
 pub use diff::DiffComponent;
+pub use filetree::FileTreeComponent;
 pub use help::HelpComponent;
 pub use msg::MsgComponent;
 pub use reset::ResetComponent;
